add js of grid banner at home
