+++ app/javascript/application.js
@@ -3,3 +3,4 @@ import "@hotwired/turbo-rails"
 import "controllers"
 import "custom/homeCarousel"
 import "custom/navLink"
+import "custom/gridBanner"
