+++ app/javascript/custom/gridBanner.js
@@ -1,0 +1,19 @@
+function initSmoothScroll() {
+  const cards = document.querySelectorAll('.js-scroll');
+  const halfWindow = window.innerHeight * 0.7;
+
+    cards.forEach((card) => {
+      const cardsTop = card.getBoundingClientRect().top;
+      const isVisible = (cardsTop - halfWindow) < 100;
+        if(isVisible) {
+          card.classList.add('js-scroll-active');
+        } else {
+          card.classList.remove('js-scroll-active');
+        }
+
+    })
+
+}
+
+window.addEventListener('scroll', initSmoothScroll);
+initSmoothScroll()
